requirements_test.txt
-r requirements_dev.txt
pytest-homeassistant-custom-component==0.13.264
#pytest>=8.3.5
#pytest-cov>=6.0.0
#pytest-asyncio>=0.26.0

